Fix M33 can't entery PD issue.
When M33 wakeup PD mode, both WUU irq and GPIO irq will give semaphore,
which causes M33 can't enter PD for second time.
This issue add a flag to indicate if a binary semaphore is taken and
blocked. If a binary semaphore is not blocked, we will not give
semaphore again.

Signed-off-by: Terry Lv <Terry.Lv@nxp.com>